fs-utils: Add ot_fd_readall_or_mmap_sized()
This is similar to ot_fd_readall_or_mmap() except that it takes both
an offset and a size, rather than always reading to the end of the
file.

Additionally this fixes a bug in ot_fd_readall_or_mmap() where the offset
has to be an exact page offset for the mmap to work.