variant-utils: Add functions for streaming reads of GVariants
The main function here is ot_variant_fd_get_child() which given a GVariant
in a file (specified by fd + offset + size) returns the offset, size and
type of a child at the specified index.

There is also a ot_variant_fd_read_child() function which combines this
with actually reading the GVariant from the fd.

This code is mainly composed of the various gvs_sometype__get_child()
functions in gvariant-serializer.c in glib.