static deltas: Do streaming reads of delta superblock and inline parts
Instead of maping the entire superblock and referencing any inline
parts from that we read the superblock in a streaming fashion from the
file descriptor, and skipping any inline delta parts. Instead we
record the offsets for said delta parts and stream them as needed from
the superblock file.

To test this i generated a flatpak bundle for the Gnome SDK, ending up
in a delta file with inlined parts at 288MB. Then i applied
this to a bare-user repo with `ostree static-delta apply-offline`
and monitored the memory size with top.

In the old code, the maximum size recorded was 556MB vss and 310MB rss.
With the streaming reads this was instead 269MB vss and 44MB rss.

This is mostly interesting for flatpak, as it uses static deltas with
inline parts for flatpak bundles.