Makefile-libostree.am: Drop Makefile dependency
For some reason, this is causing `cargo build` to execute both during a
`make` and a `make install`. This in turn is causing problems on my
system since I'm using `rustup`, so `sudo make install` fails to find
`cargo` in the more restricted `PATH`.

Looking at the librsvg code as an example, I also don't see this
dependency. It seems like any `make` command implicitly depends on
making sure the various Automake makefiles are up to date anyway?