ostree commit: Add option `--tree=prefix=PREFIX`
This enhances `ostree commit`'s ability to compose trees.  You can now run

    ostree commit -b rootfs \
        --tree=ref=rootfs \
        --tree=prefix=/usr/lib/machines/container --tree=ref=container

and that will "mount" the container ref under usr/lib/machines/container
in rootfs.  It is much faster than checking the tree out and checking it
back in again.

This is useful for our buildsystem where we ship containers in the rootfs
image.