Add private API `_ostree_repo_open_file`
`_ostree_repo_open_file` is intended to be used by our command line
utilities to convert from the text description of a commit/tree/ref, etc.
into an `OstreeRepoFile` representing it.  It takes care to give fairly
good error messages.

This commit introduces the concept of a treeish.  This is intended to make
the command line interface more flexible.  We now have a way of referring
to a directory or a file within a commit.  Once we migrate the command
line tools over to it they will become more general.

So with this we'll be able to write:

    :myref:a/path

to refer to the tree a/path under commit pointed to by myref.

`ostree_repo_read_file` is a more general version of
`ostree_repo_read_commit` and can be used in its place for greater
flexibility when you're just after a tree anyway and don't need access to
the commit object itself.

I'm implementing this because I want greater flexibility when composing
trees with `ostree commit`.