ostree diff: Allow diffing arbitrary trees and subdirs
There is a change in the output here.  Previously for files in the repo
each filename in the diff would be prefixed with /.  For files on the
filesystem there was no / prefix.  Now there is never a / prefix.  This is
more self-consistent and is the same behaviour as `git diff --stat`.