ostree ls: Accept any treeish in place of COMMIT
This is a generalisation of ostree ls to be consistent with other  places
we accept a treeish.  It doesn't behave quite as you might expect: the
printed paths are always relative to the commit root rather than to the
treeish passed in.  So:

    ostree ls :rootfs:usr/bin systemd

Prints:

    -00755 0 0  12345 /usr/bin/systemd

rather than

    -00755 0 0  12345 systemd

as you might expect.  I think it would be nice to change this behaviour,
but it would be backwards-incompatible so it can't happen.