_ostree_repo_open_file: Allow passing a DIRTREE SHA
Command-line users of `_ostree_repo_open_file` can now take a DIRTREE
SHA if they are expecting to deal with a tree object.  This is a
generalisation intended to make the command-line interface more flexible.

There is one limitation: If you pass a dirtree sha you must also pass a
subdir to use.  We need this because otherwise the tree won't have an
associated DIRMETA sha.

e.g.

    ostree ls [DIRTREE SHA] bin

would now be valid whereas

    ostree ls [DIRTREE SHA]

is not.